fixed vbsp
should be able to compile very large map now

+++ public/disp_common.cpp
@@ -616,10 +616,10 @@ static inline void SetupCoordXY( CNodeVert &out, CNodeVert const &bottomLeft, CN
 }
 
 
-static unsigned short* DispCommon_GenerateTriIndices_R(
+static int* DispCommon_GenerateTriIndices_R(
 	CNodeVert const &bottomLeft, 
 	CNodeVert const &topRight,
-	unsigned short*indices,
+	int *indices,
 	int power,
 	int sideLength )
 {
@@ -724,7 +724,7 @@ int DispCommon_GetNumTriIndices( int power )
 }
 
 
-void DispCommon_GenerateTriIndices( int power, unsigned short*indices )
+void DispCommon_GenerateTriIndices( int power, int *indices )
 {
 	int sideLength = 1 << power;
 	DispCommon_GenerateTriIndices_R(
